+++ e2e/ruby/spec/helpers.rb
@@ -268,7 +268,7 @@ module E2ERuby
       expect(elements.length).to be >= min_count if min_count
       return unless types_include
 
-      found_types = elements.filter_map(&:type).uniq
+      found_types = elements.filter_map(&:element_type).uniq
       types_include.each do |t|
         expect(found_types).to include(t)
       end